More matching and sorting options, a couple usability improvements (#62)
* Fix to allow first choice to contain leading whitespace

* Add commandline switch: auto-select single choice

After filtering the available choices with the query string, sometimes it may
turn out that only 1 choice remains. In this case, one may want choose to just
select this one choice without waiting for the user to hit Enter.

This adds a command-line switch (-1) to the code that implements this
behavior.

* Fix to not select initial query text

When we supply an initial query, due to the way macOS's text fields work, this
initial query is selected. So, any subsequent text we type ends up replacing the
initial query rather than appending to it.

This commit makes sure that the above does not happen. When an initial query is
present, any subsequent keypresses now have the effect of appending to said
query rather than replacing it.

* Add option to match words instead of characters

Previously, choose could only match each element against the query character by
character. That is, if all characters in the query appeared in the element (in
order), the element was considered a match. Even if the characters didn't appear
in sequence.

For example, the row "banana" would match the query "bn" because the string
"banana" contains the string "bn" as a sub-sequence -- even though "b" and "n"
don't appear as consecutive characters anywhere in banana.

This adds an option (-W) to only consider rows as matching if the words
in the query appear in them as *continuous* sub-strings (NOT subsequences) -- in
order.

For example, with this option enabled, "banana" would match "an" but not "bn".
And "the quick brown fox" would match "qui fox" but not "qi fx" or "fox qui".

* Add option to skip sorting matches

Previously, choose would assign a score to each matching row, and present these
rows sorted by their score.

This adds an option (-S) that when enabled, skips this scoring and
sorting. So, matches will simply be presented in whatever order they appeared
in the input.

* Support case-sensitive and smart-case matching

At the moment, choose does not allow us to match against queries
case-sensitively. It also does not allow smart-case (ie, case-sensitive when the
query contains uppercase characters, but case-insensitive if the query is
all-lowercase).

We now support these matching options. Specifically, this adds a
command-line switch (-C) that takes 1 argument:

    - '-C i' = case-insensitive (the default),
    - '-C I' = case-sensitive, and
    - '-C s' = smart-case.

Co-authored-by: Karthik V. Aadithya <kvaadit@sandia.gov>